♻️(back) use item id as label in its path
The labels used in the item path were incremented from the previous path
used. To do this, we needed to fetch the nearest sibling, increment it
to insert it. This causes race condition, hard move and hard code. We
choose to change it by using the item id as label.