🔧(back) new settings to configure boto3 checksum validation
Since version 1.36.0 the s3 client behavior has been updated to always
calculate the CRC32 checksum. This behavior is not always the default one
for solution implementing the s3 protocol. A workaround to this new
behavior is to configure the environment variables
AWS_REQUEST_CHECKSUM_CALCULATION and AWS_RESPONSE_CHECKSUM_VALIDATION
with the value `when_required`. The default value `when_supported` is
used in the settings.