♻️(back) move check constraint on filename in save method
The check constraint on the filename can be moved in the save method. A
check constraint will execute a sql query and in this case it is not
needed, we have all the element needed to do this check in the python
code and not in the database. This will reduce the number queries on
every save.