✨(front) create EmbeddedExplorer
As we started to have multiple embedded explorer in the app
it became redundant to implement the same logic again and again.
It was making DX worse. At the moment the goal is to used the
same shared component from app navigation, move modal and sdk file
picker.
This component is intended to be completelly standalone, but it
is not the case yet, we still need to make it independant from
GlobalExplorerProvider and TreeProvider some day. For now it does
the job well avoiding logic duplication.