🐛(global) fix wrong Content-Type on specific s3 implementations
When requesting file from the Scaleway object storage the
Content-Type in the response header contain value like
multipart/form-data; boundary=----WebKitFormBoundaryP....
instead of expected values containing mime type like application/pdf.

This was causing issue when previewing pdf, it was triggering
a download instead of the browser native preview.

This was only happening on Scaleway object storage, it is
totally fine with MinIO which seems to guess itself the
Content-Type based on Claude conversation.