♻️(front) move tree building into TreeProvider
We want to be able to use getAncestors method from treeContext even
when the tree is hidden. Like when visiting as anon, the left panel
is not mounted, and because the tree was built from the not rendered
ExplorerTree, we were not able to build the breadcrumbs in those
cases. Also make some cleanup and introduce spinner during the
first load of GlobalExplorerContext.