🐛(backend) stop decreasing twice the numchild in deletion process
In the deletion process we maintain the numchild property by decreasing
when the item is soft deleted. But the delete method is also decreasing
it. We must only rely on the soft delete process and ensure that an item
can not be deleted if it was not solft_deleted first.
Fixes #261