♻️(back) handle specific allow cors for sdk relay
We want to be able to specify which origins are allowed to pull the
event route. This way we introduce SDK_ALLOWED_ORIGINS.

I did not found a clean and easy to maintain way to to do the same
behavior with django-cors-headers which aim to provide a global
behavior accross the API, here we want a special behavior only for
the sdk relay routes.

Also remove the possibility to use the POST route from cross
domain.