🧑‍💻(keycloak) do not require ssl for dev server
Set `sslRequired` to `none` for all realms in the keycloak dev service.
By default this settings is set to `external` and only local IPs are granted
to access to the server without SSL. But sometimes, with docker (desktop?)
it appears the network configuration may keycloak to consider the accessing
ip as not local.

We also took opportunity of this change to use the latest keycloak image version

Resolve #291