♻️(back) manage title uniqueness by generating new one if existing
For now it wasn't possible to create twice an item with the same title
in the same path and an exception was raised. We change it by
incrementing a version number in the item title. It is still not
possible to have twice the same title but the item is created with a
different one.