🔧(conf) setup handy local e2e env
We want to be able to quickly start a local isolated env for the
e2e testing, we don't want to share to same db with the classic
dev env because the e2e db should be empty before tests execution.
And we don't want to clear our dev db every time. Also provide a new
makefile handy command to launch the e2e env and the tests in one
command, great.