✨(back) add demo app with quick authent
We want to be able to login users quickly from the e2e frontend tests.
Because sometimes we may need to reset the DB to run the test, instead
of spending 5sec of going through the login process via keycloak, just
calling this new api route cost 50ms, and voilà. Really great for DX
and tests performances.