♻️(front) externalize FilePreview from EmbeddedExplorer
Apart from the paradigm of keeping the EmbeddedExplorer a black box that
ONLY displays a file explorer, without providing advanced peripheric
features, I faced a dead-end when implementing the opening of the FilePreview
from the search modal. Indeed the search modal is outside of the
EmbeddedExplorer, so that lead to ... how do I open a FilePreview from
search? By removing it from the EmbeddedExplorer and making it external.