♻️(front) modalify and fix bug of FilePreview
The initial logic was always triggering onChangeFile containing the
first item of the list on first load, even if openedFileId is null and
the FilePreview closed, that was causing infinite rendering loop with
the new imlementation of it inside the GlobalExplorerContext.
Also make this component use the modals, because we want to be able to
render modals in the order on which they are displayed. Before the
FilePreview was hard-coded to always display behind the modals in order
to make visible confirmation modals when clicking download. But we want
the opposite when opening a preview from the search modal, we want the
search modal to stay behind.