♻️(front) refactor tree loading
When using the search modal to teleport to a found folder that is not yet
loaded in the left tree, we were unable to render the breadcrumbs and to
update the tree. As of now, it was always lazy loaded. So by now loading
the /tree from the AppExplorerGrid, we can ensure we always load it if
needed, it also simplify the logic inside TreeInitializer that was handling
too much, like edge cases were prevent loading /tree if we were on a trash
route, etc ... Logic inversion stuff, you got it.