✨(backend) force content-disposition header
For the files without the preview feature, we want to force the
content-disposition to attachment. We guess the mimetype based on the
file extension and if it is not in the previewable list, we set the
ContentDisposition to the upload policy. When we will fetch the url,
this header will be added.