✨(front) add creating folder upload step
Before when starting an upload the toast was not shown
during the creation phase of folders. It was not visible
when uploading small folders trees because it was quick.
But when uploading a huge folders tree the interface
was not giving any feedback during dozens of seconds while
creating folders in background. Not we want to immediately
show the toast with a specific message.