✨(backend) add is_explicit field on ItemAccessSerializer
We need to know, for a given if the related access is explicit or
inherited. This is usefull on the list endpoint. This endpoint returns
the entire list of accesses add on the entire tree, we need to know when
retrieving this list if each access is explicit on the item or
inherited.