🐛(wopi) handle rename file encoded in utf7
The wopi spec indicates that the new filename requested when a file is
renamed is encoded using utf-7. This lead to filename containing weird
characters in their filename when not managed correctly. Also, to be
consistent with the core api, the title is also changed using the
requested filename without the extension