✨(back) add entitlement route
The frontend needs to be able to get the entitlements
values before performing actions knowing they will fail.
The exemple is a mass upload of dozens files and folders,
we want to know before creating folders and starting uploads
that it will not be allowed anyway. This provides a better
UX.