🐛(backend) manage file renaming when filename has not changed
The rename feature does not check if the filename is effectly changed
before renaming the file on the storage. This behavior leads to an error
because it is not alloed to copy an object using the same name without
modifying its metadata. We now check if the filename is hcanged.