🐛(backend) managed empty uploaded files
When an empty file is uploaded, or a file with a size less than 2048
bytes, the uploaded_ended endpoint is failing. The requested range in
the http request can't be higher than the file size itself. To fix it we
retrieve first file metadata and check its size in the given metadata to
adapt the range.