🐛(backend) stop renaming file when no title is provided
In the ItemViewset::perfom_update method, we trigger the rename_file
task when the title of the item has changed. On a PATCH request, the
title can be absent in the payload leading to renaming the file with a
None value. We have to chank first of the title is present in the
validated_data and also check first in the renamin task if the provided
value is not empty or None.
Fix # 477